Refactor the facility scrollable list
